применена деструктуризация к компоненту profile
