+++ src/components/Dialogs/DialogItem/DialogItem.module.css
@@ -1,0 +1,23 @@
+.dialogs {
+    display: grid;
+    grid-template-columns: 2fr 10fr;
+}
+
+.dialogItems {
+    padding: 10px;
+
+}
+
+.dialogItems .active {
+    color: brown;
+
+}
+
+.dialog {
+    padding: 10px;
+}
+
+.messages {
+    padding: 10px;
+}
+
