+++ src/components/Dialogs/Dialogs.jsx
@@ -4,26 +4,9 @@ import { NavLink } from "react-router-dom";
 import { ROUTES, I18N } from "../../core/constants";
 import ACTION_TYPES from "../../redux/actionTypes";
 
+import {dialogsElements, messagesElements} from "./DialogItem/DialogItem"
+
 import s from "./Dialogs.module.css";
-
-const DialogItem = (props) => {
-  const way = `${ROUTES.MESSAGES}/${props.id}`;
-  return (
-    <div className={`${s.dialog} ${s.active}`}>
-      <NavLink to={way}>{props.name}</NavLink>
-    </div>
-  );
-};
-
-const Message = (props) => {
-  return <div className={s.dialogs}>{props.message}</div>;
-};
-
-const dialogsElements = (usersData) =>
-  usersData.map((dialog) => <DialogItem name={dialog.name} id={dialog.id} />);
-
-const messagesElements = (messagesData) =>
-  messagesData.map((message) => <Message message={message.message} />);
 
 const Dialogs = (props) => {
   const {dialogsPage, dispatch} = props;
